test(menu): share menu context helper
